Arrays Lab
In this lab, we created a variable cats and assigned it to an array with three cat names inside. We then created 8 functions. The first for modify the array with .push(), .unshift, .pop and .shift. The last four functions create a copy of the array. The next two out of the four add a cat to the copy one at the end and one at  the begining using the spreat[...] operator. The last two use the .slice method to remove the last cat and the first cat from the copy of the array.